requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==4.0
django-tastypie==0.14.4
kombu==5.2.4
mysqlclient==2.1.1
prompt-toolkit==3.0.36
python-dateutil==2.8.2
python-mimeparse==1.6.0
python3-digest==1.8b4
pytz==2022.7
redis==4.4.0
six==1.16.0
sqlparse==0.4.3
structlog==22.3.0
tzdata==2022.7
vine==5.0.0
wcwidth==0.2.5
